prevented post process texture sampler from wrapping

+++ DX11Framework/FLight.cpp
@@ -11,6 +11,6 @@ void FLight::convertToData(FLightData* data)
 	XMFLOAT4X4 world_mat = getTransform();
 	XMFLOAT3 direction = XMFLOAT3(-world_mat._31, -world_mat._32, -world_mat._33);
 
-	data->light_direction = XMFLOAT4(direction.x, direction.y, direction.z, type == FLightType::DIRECTIONAL ? 0 : 1);
+	data->light_direction = XMFLOAT4(direction.x, direction.y, direction.z, (float)(type == FLightType::DIRECTIONAL ? 0 : 1));
 	data->light_position = XMFLOAT3(world_mat._41, world_mat._42, world_mat._43);
 }
